+++ src/app/components/index.ts
@@ -1,4 +1,4 @@
-export { PageViewer } from './Layout';
+export { PageView } from './Layout';
 export { PageViewComponent} from './PageViewComponent';
 export { SeriesComponent } from './SeriesComponent';
 export { ComicsComponent } from './ComicsComponent';
