+++ src/app/models/konva.models.ts
@@ -2,7 +2,3 @@ export interface PositionOnLayer {
   x: number;
   y: number;
 }
-
-export interface PageTiles {
-  Array<>
-}
